+++ src/components/src/Item.cpp
@@ -13,7 +13,7 @@ Item::Item(YAML::Node data, AssetManager* assets)
     observations_ = data["observations"].as<std::vector<std::string>>();
     pick_observation_ = data["pick_observation"].as<std::string>();
     // Load click area polygon
-    click_area_.add_vertices(data["click_area"].as<std::vector<std::vector<int>>>());
+    click_area_.add_vertices(data["click_area"].as<std::vector<std::vector<float>>>());
     // Load texture
     sprite_ = assets->sprite(id_);  // TO DO: log error if assets return nullptr
 }
@@ -30,7 +30,7 @@ void Item::update(RenderManager* renderer, int dt)
 }
 
 
-bool Item::clicked(int x, int y)
+bool Item::clicked(float x, float y)
 {
     return click_area_.point_in_polygon(x, y) && state_;
 }
@@ -48,26 +48,26 @@ std::string Item::pick_observation()
 }
 
 
-bool Item::lock() { return lock_; }
-void Item::lock(bool new_lock) { lock_ = new_lock; }
-Sprite* Item::sprite() { return sprite_; }
-Polygon* Item::click_area() { return &click_area_; }
+bool Item::lock()              { return lock_;        }
+void Item::lock(bool new_lock) { lock_ = new_lock;    }
+Sprite* Item::sprite()         { return sprite_;      }
+Polygon* Item::click_area()    { return &click_area_; }
 
-void Item::x(int x)
+void Item::x(float x)
 {
-    int old_x { sprite_->x() };
-    int dx { x - old_x };
+    float old_x { sprite_->x() };
+    float dx { x - old_x };
     click_area_.move(dx, 0);
     sprite_->x(x);
 }
-void Item::y(int y)
+void Item::y(float y)
 {
-    int old_y { sprite_->y() };
-    int dy { y - old_y };
+    float old_y { sprite_->y() };
+    float dy { y - old_y };
     click_area_.move(0, dy);
     sprite_->y(y);
 }
-void Item::move (int dx, int dy)
+void Item::move (float dx, float dy)
 {
     sprite_->move(dx, dy);
     click_area_.move(dx, dy);
